Extract duplicate code into common baseclass
