+++ app/enricher/prepare_state.py
@@ -2,47 +2,41 @@
 Provides enricher strategy for enriching :class:`~app.model.CompileRequest.PrepareStateNode` from a database.
 """
 
-from typing import override
+from typing import cast, override
 
-from sqlalchemy import exists, select
-from sqlalchemy.ext.asyncio import AsyncEngine, AsyncSession
+from sqlalchemy import Select, exists, select
+from sqlalchemy.ext.asyncio import AsyncEngine
 
 from app.enricher import (
     Constraints,
     ConstraintValidationException,
-    EnricherStrategy,
-    EnrichmentResult,
-    ImplementationMetaData,
     InputValidationException,
-    NodeUnsupportedException,
 )
-from app.enricher.models import Input, NodeType, QuantumStateType
+from app.enricher.db_enricher import DataBaseEnricherStrategy
+from app.enricher.models import BaseNode, Input, NodeType, QuantumStateType
 from app.enricher.models import PrepareStateNode as PrepareStateTable
-from app.model.CompileRequest import (
-    ImplementationNode,
-    PrepareStateNode,
-)
 from app.model.CompileRequest import (
     Node as FrontendNode,
 )
+from app.model.CompileRequest import (
+    PrepareStateNode,
+)
 
 
-class PrepareStateEnricherStrategy(EnricherStrategy):
+class PrepareStateEnricherStrategy(DataBaseEnricherStrategy):
     """
     Strategy capable of enriching :class:`~app.model.CompileRequest.PrepareStateNode` from a database.
     """
 
-    engine: AsyncEngine
-
     def __init__(self, engine: AsyncEngine):
-        self.engine = engine
+        super().__init__(engine)
 
     @override
-    async def _enrich_impl(
+    def _generate_query(
         self, node: FrontendNode, constraints: Constraints | None
-    ) -> list[EnrichmentResult]:
+    ) -> Select[tuple[BaseNode]] | None:
         if not isinstance(node, PrepareStateNode):
-            return []
+            return None
 
         if node.quantumState == "custom" or node.size <= 0:
             raise InputValidationException(
@@ -53,33 +47,12 @@ class PrepareStateEnricherStrategy(EnricherStrategy):
             raise ConstraintValidationException("PrepareStateNode can't have an input")
 
         no_inputs = ~exists().where(Input.node_id == PrepareStateTable.id)
-        query = select(PrepareStateTable).where(
-            PrepareStateTable.type == NodeType(node.type),
-            no_inputs,
-            PrepareStateTable.quantum_state == QuantumStateType(node.quantumState),
-            PrepareStateTable.size == node.size,
+        return cast(
+            Select[tuple[BaseNode]],
+            select(PrepareStateTable).where(
+                PrepareStateTable.type == NodeType(node.type),
+                no_inputs,
+                PrepareStateTable.quantum_state == QuantumStateType(node.quantumState),
+                PrepareStateTable.size == node.size,
+            ),
         )
-
-        async with AsyncSession(self.engine) as session:
-            result_nodes = (await session.execute(query)).scalars().all()
-
-        if not result_nodes:
-            raise RuntimeError("No results found in the database")
-
-        enrichment_results = []
-        for result_node in result_nodes:
-            if result_node.implementation is None:
-                raise NodeUnsupportedException(node)
-
-            enrichment_results.append(
-                EnrichmentResult(
-                    ImplementationNode(
-                        id=node.id, implementation=result_node.implementation
-                    ),
-                    ImplementationMetaData(
-                        width=result_node.width, depth=result_node.depth
-                    ),
-                )
-            )
-
-        return enrichment_results
